Move Splash kernel planning into shared helper
