+++ lib/levanter/src/levanter/layers/attention.py
@@ -8,24 +8,24 @@ import math
 import warnings
 from dataclasses import dataclass
 from enum import StrEnum
-from typing import Literal, Optional, TypedDict, Union, cast
+from typing import Literal, Optional, TypedDict, Union
 
 import equinox as eqx
 import jax
 import jax.random as jrandom
 from jax import numpy as jnp
-from jax.experimental.pallas.ops.tpu.splash_attention import splash_attention_kernel
 
 from levanter.kernels.pallas.splash_attention import (
     DEFAULT_SPLASH_BLOCK_SIZE,
     SPLASH_BLOCK_GRANULARITY,
+    SplashKernelContext,
+    SplashKernelPlan,
+    SplashMaskPlanSpec,
+    SplashPrefixControls,
+    SplashSegmentRunControls,
     SplashSegmentIdsLowering,
-    lower_splash_attention_mask,
     lower_splash_segment_ids,
-    packed_causal_segment_mask_infos,
-    packed_causal_segment_run_mask_infos,
-    packed_prefix_lm_segment_run_mask_infos,
-    prefix_lm_mask_infos,
+    prepare_splash_kernel_plan,
     splash_attention_block_sizes,
     splash_attention_mask_spec_from_fields,
     splash_partition_spec_shard_factor,
@@ -885,79 +885,6 @@ def _prepare_splash_batch_value(
     return _maybe_flatten(value, (), SPLASH_BATCH_AXIS_NAME)
 
 
-def _batched_splash_kernel_specs(
-    kernel,
-    *,
-    batch_spec,
-    head_spec,
-    q_seq_spec,
-    num_heads: int,
-):
-    def spec_for_leaf(leaf):
-        if leaf is None:
-            return None
-        if not isinstance(leaf, jax.Array):
-            return PartitionSpec()
-        if leaf.ndim == 0:
-            return PartitionSpec()
-        if leaf.ndim >= 4 and leaf.shape[1] == num_heads:
-            return PartitionSpec(batch_spec, head_spec, q_seq_spec, *([None] * (leaf.ndim - 3)))
-        return PartitionSpec(batch_spec, *([None] * (leaf.ndim - 1)))
-
-    return jax.tree_util.tree_map(spec_for_leaf, kernel)
-
-
-def _splash_kernel_from_dynamic_metadata(metadata, block_sizes, logits_soft_cap):
-    return splash_attention_kernel.SplashAttentionKernel(
-        metadata.fwd_mask_info,
-        None if block_sizes.use_fused_bwd_kernel else metadata.dq_mask_info,
-        metadata.dkv_mask_info,
-        block_sizes=block_sizes,
-        is_mqa=False,
-        save_residuals=False,
-        mask_value=splash_attention_kernel.DEFAULT_MASK_VALUE,
-        attn_logits_soft_cap=logits_soft_cap,
-        residual_checkpoint_name=None,
-        mask_function=None,
-        interpret=False,
-    )
-
-
-@dataclass(frozen=True)
-class _SplashPrefixControls:
-    prefix_lengths: jax.Array | None
-    physical_axes_prefix_lengths: PartitionSpec | None
-    prefix_lengths_per_segment: jax.Array | None
-    physical_axes_prefix_lengths_per_segment: PartitionSpec | None
-
-
-@dataclass(frozen=True)
-class _SplashSegmentRunControls:
-    segment_lengths: jax.Array | None
-    physical_axes_segment_lengths: PartitionSpec | None
-    num_segments: jax.Array | None
-
-
-@dataclass(frozen=True)
-class _SplashKernelPlan:
-    kernel: object
-    kernel_specs: object
-    kernel_vmap_axis: int | None
-    segment_id_lowering: SplashSegmentIdsLowering
-
-
-@dataclass(frozen=True)
-class _SplashKernelContext:
-    q_seq_len: int
-    kv_seq_len: int
-    num_heads: int
-    block_sizes: splash_attention_kernel.BlockSizes
-    head_shards: int
-    q_seq_shards: int
-    physical_axes_q: PartitionSpec
-    logits_soft_cap: float | None
-
-
 @dataclass(frozen=True)
 class _SplashPreparedLayout:
     q: jax.Array
@@ -980,17 +907,9 @@ class _SplashPreparedLayout:
 @dataclass(frozen=True)
 class _SplashInvocationPlan:
     mesh: Mesh
-    kernel_plan: _SplashKernelPlan
+    kernel_plan: SplashKernelPlan
     sinks: jax.Array | None
     physical_axes_sink: PartitionSpec | None
-
-
-@dataclass(frozen=True)
-class _PackedSegmentIds:
-    q: jax.Array
-    kv: jax.Array
-    q_vmap_axis: int | None
-    kv_vmap_axis: int | None
 
 
 def _flatten_partition_spec_entries(entries):
@@ -1026,7 +945,7 @@ def _prepare_splash_prefix_controls(
     mask: Optional[Union[NamedArray, "AttentionMask"]],
     q_class,
     physical_axes_q: PartitionSpec,
-) -> _SplashPrefixControls:
+) -> SplashPrefixControls:
     prefix_lengths = None
     physical_axes_prefix_lengths = None
     prefix_lengths_per_segment = None
@@ -1068,7 +987,7 @@ def _prepare_splash_prefix_controls(
         if prefix_lm.prefix_length is not None:
             prefix_lengths = jnp.maximum(prefix_lengths, prefix_lm.prefix_length)
 
-    return _SplashPrefixControls(
+    return SplashPrefixControls(
         prefix_lengths=prefix_lengths,
         physical_axes_prefix_lengths=physical_axes_prefix_lengths,
         prefix_lengths_per_segment=prefix_lengths_per_segment,
@@ -1081,9 +1000,9 @@ def _prepare_splash_segment_run_controls(
     mask: Optional[Union[NamedArray, "AttentionMask"]],
     q_class,
     physical_axes_q: PartitionSpec,
-) -> _SplashSegmentRunControls:
+) -> SplashSegmentRunControls:
     if not isinstance(mask, AttentionMask) or mask.segment_run_metadata is None:
-        return _SplashSegmentRunControls(
+        return SplashSegmentRunControls(
             segment_lengths=None,
             physical_axes_segment_lengths=None,
             num_segments=None,
@@ -1111,7 +1030,7 @@ def _prepare_splash_segment_run_controls(
     )
     num_segments = num_segments.rearrange((SPLASH_BATCH_AXIS_NAME,))
 
-    return _SplashSegmentRunControls(
+    return SplashSegmentRunControls(
         segment_lengths=segment_lengths.astype(jnp.int32).array,
         physical_axes_segment_lengths=PartitionSpec(physical_axes_q[0], None),
         num_segments=num_segments.astype(jnp.int32).array,
@@ -1144,403 +1063,25 @@ def _lower_splash_attention_segment_ids(
     )
 
 
-def _segment_ids_for_packed_mask(
-    segment_id_lowering: SplashSegmentIdsLowering,
-    *,
-    q_seq_len: int,
-    kv_seq_len: int,
-) -> _PackedSegmentIds:
-    if segment_id_lowering.segment_ids is None:
-        q_segment_ids = jnp.zeros((q_seq_len,), dtype=jnp.int32)
-        kv_segment_ids = jnp.zeros((kv_seq_len,), dtype=jnp.int32)
-        return _PackedSegmentIds(q=q_segment_ids, kv=kv_segment_ids, q_vmap_axis=None, kv_vmap_axis=None)
-
-    assert segment_id_lowering.segment_batch_axis is not None
-    return _PackedSegmentIds(
-        q=segment_id_lowering.segment_ids.q,
-        kv=segment_id_lowering.segment_ids.kv,
-        q_vmap_axis=cast(int | None, segment_id_lowering.segment_batch_axis.q),
-        kv_vmap_axis=cast(int | None, segment_id_lowering.segment_batch_axis.kv),
-    )
-
-
-def _packed_dynamic_mask_kernel_plan(
-    *,
-    segment_id_lowering: SplashSegmentIdsLowering,
-    context: _SplashKernelContext,
-    metadata_values: tuple[jax.Array, ...],
-    metadata_in_axes: tuple[int | None, ...],
-    metadata_builder,
-    batch_spec,
-) -> _SplashKernelPlan:
-    segment_ids = _segment_ids_for_packed_mask(
-        segment_id_lowering,
-        q_seq_len=context.q_seq_len,
-        kv_seq_len=context.kv_seq_len,
-    )
-    if batch_spec is None:
-        assert segment_id_lowering.segment_ids_axes is not None
-        assert segment_ids.q_vmap_axis is not None
-        batch_spec = segment_id_lowering.segment_ids_axes.q[segment_ids.q_vmap_axis]
-
-    def make_kernel(*args):
-        *metadata_args, q_segment_ids, kv_segment_ids = args
-        return metadata_builder(*metadata_args, q_segment_ids, kv_segment_ids)
-
-    return _dynamic_metadata_kernel_plan(
-        context=context,
-        metadata_values=metadata_values + (segment_ids.q, segment_ids.kv),
-        metadata_in_axes=metadata_in_axes + (segment_ids.q_vmap_axis, segment_ids.kv_vmap_axis),
-        metadata_builder=make_kernel,
-        batch_spec=batch_spec,
-        segment_id_lowering=lower_splash_segment_ids(),
-    )
-
-
-def _dynamic_metadata_kernel_plan(
-    *,
-    context: _SplashKernelContext,
-    metadata_values: tuple[jax.Array, ...],
-    metadata_in_axes: tuple[int | None, ...],
-    metadata_builder,
-    batch_spec,
-    segment_id_lowering: SplashSegmentIdsLowering,
-) -> _SplashKernelPlan:
-    def make_kernel(*metadata_args):
-        metadata = metadata_builder(*metadata_args)
-        return _splash_kernel_from_dynamic_metadata(metadata, context.block_sizes, context.logits_soft_cap)
-
-    splash_kernel = jax.vmap(make_kernel, in_axes=metadata_in_axes)(*metadata_values)
-    kernel_specs = _batched_splash_kernel_specs(
-        splash_kernel,
-        batch_spec=batch_spec,
-        head_spec=context.physical_axes_q[1],
-        q_seq_spec=context.physical_axes_q[2],
-        num_heads=context.num_heads,
-    )
-    return _SplashKernelPlan(
-        kernel=splash_kernel,
-        kernel_specs=kernel_specs,
-        kernel_vmap_axis=0,
-        segment_id_lowering=segment_id_lowering,
-    )
-
-
-def _can_use_packed_causal_segment_kernel(
-    *,
-    mask: Optional[Union[NamedArray, "AttentionMask"]],
-    prefix_controls: _SplashPrefixControls,
-    segment_id_lowering: SplashSegmentIdsLowering,
-    context: _SplashKernelContext,
-) -> bool:
-    if not isinstance(mask, AttentionMask):
-        return False
-    if context.q_seq_shards != 1:
-        return False
-    if segment_id_lowering.segment_ids is None or segment_id_lowering.segment_batch_axis is None:
-        return False
-    if segment_id_lowering.segment_batch_axis.q is None or segment_id_lowering.segment_batch_axis.kv is None:
-        return False
-    return _is_plain_causal_splash_mask(mask=mask, prefix_controls=prefix_controls)
-
-
-def _is_plain_causal_splash_mask(
-    *,
-    mask: Optional[Union[NamedArray, "AttentionMask"]],
-    prefix_controls: _SplashPrefixControls,
-) -> bool:
-    return (
-        isinstance(mask, AttentionMask)
-        and mask.is_causal
-        and mask.causal_offset is None
-        and mask.sliding_window is None
-        and mask.prefix_lm_spec is None
-        and mask.explicit_mask is None
-        and prefix_controls.prefix_lengths is None
-        and prefix_controls.prefix_lengths_per_segment is None
-    )
-
-
-def _can_use_packed_prefix_lm_segment_run_kernel(
-    *,
-    mask: Optional[Union[NamedArray, "AttentionMask"]],
-    prefix_controls: _SplashPrefixControls,
-    segment_run_controls: _SplashSegmentRunControls,
-    context: _SplashKernelContext,
-) -> bool:
-    if not isinstance(mask, AttentionMask):
-        return False
-    if context.q_seq_shards != 1:
-        return False
-    if context.q_seq_len != context.kv_seq_len:
-        return False
-    if segment_run_controls.segment_lengths is None or segment_run_controls.num_segments is None:
-        return False
-    prefix_lm = mask.prefix_lm_spec
-    return (
-        mask.is_causal
-        and mask.causal_offset is None
-        and mask.sliding_window is None
-        and mask.explicit_mask is None
-        and prefix_lm is not None
-        and prefix_lm.prefix_length is None
-        and prefix_lm.prefix_lengths is None
-        and prefix_controls.prefix_lengths_per_segment is not None
-    )
-
-
-def _can_use_packed_causal_segment_run_kernel(
-    *,
-    mask: Optional[Union[NamedArray, "AttentionMask"]],
-    prefix_controls: _SplashPrefixControls,
-    segment_run_controls: _SplashSegmentRunControls,
-    context: _SplashKernelContext,
-) -> bool:
-    if not isinstance(mask, AttentionMask):
-        return False
-    if context.q_seq_shards != 1:
-        return False
-    if context.q_seq_len != context.kv_seq_len:
-        return False
-    if segment_run_controls.segment_lengths is None or segment_run_controls.num_segments is None:
-        return False
-    return _is_plain_causal_splash_mask(mask=mask, prefix_controls=prefix_controls)
-
-
-def _packed_causal_segment_run_kernel_plan(
-    *,
-    segment_run_controls: _SplashSegmentRunControls,
-    context: _SplashKernelContext,
-) -> _SplashKernelPlan:
-    assert segment_run_controls.segment_lengths is not None
-    assert segment_run_controls.num_segments is not None
-    assert segment_run_controls.physical_axes_segment_lengths is not None
-
-    def make_kernel(segment_lengths, num_segments):
-        metadata = packed_causal_segment_run_mask_infos(
-            segment_lengths=segment_lengths,
-            num_segments=num_segments,
-            q_seq_len=context.q_seq_len,
-            kv_seq_len=context.kv_seq_len,
-            block_sizes=context.block_sizes,
-            head_shards=context.head_shards,
-            q_seq_shards=context.q_seq_shards,
-        )
-        return _splash_kernel_from_dynamic_metadata(metadata, context.block_sizes, context.logits_soft_cap)
-
-    return _dynamic_metadata_kernel_plan(
-        context=context,
-        metadata_values=(segment_run_controls.segment_lengths, segment_run_controls.num_segments),
-        metadata_in_axes=(0, 0),
-        metadata_builder=make_kernel,
-        batch_spec=segment_run_controls.physical_axes_segment_lengths[0],
-        segment_id_lowering=lower_splash_segment_ids(),
-    )
-
-
-def _packed_prefix_lm_segment_run_kernel_plan(
-    *,
-    prefix_controls: _SplashPrefixControls,
-    segment_run_controls: _SplashSegmentRunControls,
-    context: _SplashKernelContext,
-) -> _SplashKernelPlan:
-    assert prefix_controls.prefix_lengths_per_segment is not None
-    assert prefix_controls.physical_axes_prefix_lengths_per_segment is not None
-    assert segment_run_controls.segment_lengths is not None
-    assert segment_run_controls.num_segments is not None
-
-    def make_kernel(prefix_lengths_per_segment, segment_lengths, num_segments):
-        metadata = packed_prefix_lm_segment_run_mask_infos(
-            prefix_lengths=prefix_lengths_per_segment,
-            segment_lengths=segment_lengths,
-            num_segments=num_segments,
-            q_seq_len=context.q_seq_len,
-            kv_seq_len=context.kv_seq_len,
-            block_sizes=context.block_sizes,
-            head_shards=context.head_shards,
-            q_seq_shards=context.q_seq_shards,
-        )
-        return _splash_kernel_from_dynamic_metadata(metadata, context.block_sizes, context.logits_soft_cap)
-
-    return _dynamic_metadata_kernel_plan(
-        context=context,
-        metadata_values=(
-            prefix_controls.prefix_lengths_per_segment,
-            segment_run_controls.segment_lengths,
-            segment_run_controls.num_segments,
-        ),
-        metadata_in_axes=(0, 0, 0),
-        metadata_builder=make_kernel,
-        batch_spec=prefix_controls.physical_axes_prefix_lengths_per_segment[0],
-        segment_id_lowering=lower_splash_segment_ids(),
-    )
-
-
-def _packed_causal_segment_kernel_plan(
-    *,
-    segment_id_lowering: SplashSegmentIdsLowering,
-    context: _SplashKernelContext,
-) -> _SplashKernelPlan:
-    def causal_metadata(q_segment_ids, kv_segment_ids):
-        return packed_causal_segment_mask_infos(
-            q_segment_ids=q_segment_ids,
-            kv_segment_ids=kv_segment_ids,
-            q_seq_len=context.q_seq_len,
-            kv_seq_len=context.kv_seq_len,
-            block_sizes=context.block_sizes,
-            head_shards=context.head_shards,
-            q_seq_shards=context.q_seq_shards,
-        )
-
-    return _packed_dynamic_mask_kernel_plan(
-        segment_id_lowering=segment_id_lowering,
-        context=context,
-        metadata_values=(),
-        metadata_in_axes=(),
-        metadata_builder=causal_metadata,
-        batch_spec=None,
-    )
-
-
-def _prefix_length_kernel_plan(
-    *,
-    prefix_lengths: jax.Array,
-    physical_axes_prefix_lengths: PartitionSpec,
-    segment_id_lowering: SplashSegmentIdsLowering,
-    context: _SplashKernelContext,
-) -> _SplashKernelPlan:
-    def make_prefix_lm_kernel(prefix_length):
-        prefix_metadata = prefix_lm_mask_infos(
-            prefix_length=prefix_length,
-            q_seq_len=context.q_seq_len,
-            kv_seq_len=context.kv_seq_len,
-            num_heads=context.num_heads,
-            block_sizes=context.block_sizes,
-        )
-        return _splash_kernel_from_dynamic_metadata(prefix_metadata, context.block_sizes, context.logits_soft_cap)
-
-    return _dynamic_metadata_kernel_plan(
-        context=context,
-        metadata_values=(prefix_lengths,),
-        metadata_in_axes=(0,),
-        metadata_builder=make_prefix_lm_kernel,
-        batch_spec=physical_axes_prefix_lengths[0],
-        segment_id_lowering=segment_id_lowering,
-    )
-
-
-def _static_splash_kernel_plan(
-    *,
-    mask: Optional[Union[NamedArray, "AttentionMask"]],
-    segment_id_lowering: SplashSegmentIdsLowering,
-    context: _SplashKernelContext,
-    mesh,
-) -> _SplashKernelPlan:
+def _splash_mask_plan_spec_from_mask(mask: Optional[Union[NamedArray, "AttentionMask"]]) -> SplashMaskPlanSpec:
     if mask is None:
-        mask_spec = None
-    elif isinstance(mask, AttentionMask):
-        mask_spec = splash_attention_mask_spec_from_fields(
-            is_causal=mask.is_causal,
-            causal_offset=mask.causal_offset,
-            sliding_window=mask.sliding_window,
-            prefix_length=None if mask.prefix_lm_spec is None else mask.prefix_lm_spec.prefix_length,
-            prefix_lengths=None if mask.prefix_lm_spec is None else mask.prefix_lm_spec.prefix_lengths,
-            explicit_mask=mask.explicit_mask,
-        )
-    elif isinstance(mask, NamedArray):
+        return SplashMaskPlanSpec(mask=None)
+    if isinstance(mask, NamedArray):
         raise NotImplementedError("NamedArray masks are not yet supported for splash attention")
-    else:
+    if not isinstance(mask, AttentionMask):
         raise ValueError(f"Unknown mask type: {mask}")
 
-    mask_lowering = lower_splash_attention_mask(
-        mask=mask_spec,
-        q_seq_len=context.q_seq_len,
-        kv_seq_len=context.kv_seq_len,
-        num_heads=context.num_heads,
-        q_seq_shards=context.q_seq_shards,
+    prefix_lm = mask.prefix_lm_spec
+    mask_spec = splash_attention_mask_spec_from_fields(
+        is_causal=mask.is_causal,
+        causal_offset=mask.causal_offset,
+        sliding_window=mask.sliding_window,
+        prefix_length=None if prefix_lm is None else prefix_lm.prefix_length,
+        prefix_lengths=None if prefix_lm is None else prefix_lm.prefix_lengths,
+        prefix_lengths_per_segment=None if prefix_lm is None else prefix_lm.prefix_lengths_per_segment,
+        explicit_mask=mask.explicit_mask,
     )
-
-    splash_kernel = splash_attention_kernel.make_splash_mha(
-        mask=mask_lowering.kernel_mask,
-        head_shards=context.head_shards,
-        q_seq_shards=context.q_seq_shards,
-        block_sizes=context.block_sizes,
-        attn_logits_soft_cap=context.logits_soft_cap,
-    )
-
-    kernel_sharding = jax.sharding.NamedSharding(
-        mesh, PartitionSpec(context.physical_axes_q[1], context.physical_axes_q[2])
-    )
-    kernel_specs = splash_kernel.manual_sharding_spec(kernel_sharding)
-    return _SplashKernelPlan(
-        kernel=splash_kernel,
-        kernel_specs=kernel_specs,
-        kernel_vmap_axis=None,
-        segment_id_lowering=segment_id_lowering,
-    )
-
-
-def _splash_kernel_plan(
-    *,
-    mask: Optional[Union[NamedArray, "AttentionMask"]],
-    prefix_controls: _SplashPrefixControls,
-    segment_run_controls: _SplashSegmentRunControls,
-    segment_id_lowering: SplashSegmentIdsLowering,
-    context: _SplashKernelContext,
-    mesh,
-) -> _SplashKernelPlan:
-    if _can_use_packed_prefix_lm_segment_run_kernel(
-        mask=mask,
-        prefix_controls=prefix_controls,
-        segment_run_controls=segment_run_controls,
-        context=context,
-    ):
-        return _packed_prefix_lm_segment_run_kernel_plan(
-            prefix_controls=prefix_controls,
-            segment_run_controls=segment_run_controls,
-            context=context,
-        )
-
-    if prefix_controls.prefix_lengths_per_segment is not None:
-        raise NotImplementedError("Splash attention could not use the packed prefix-LM segment-run fast path.")
-
-    if prefix_controls.prefix_lengths is not None:
-        assert prefix_controls.physical_axes_prefix_lengths is not None
-        return _prefix_length_kernel_plan(
-            prefix_lengths=prefix_controls.prefix_lengths,
-            physical_axes_prefix_lengths=prefix_controls.physical_axes_prefix_lengths,
-            segment_id_lowering=segment_id_lowering,
-            context=context,
-        )
-
-    if _can_use_packed_causal_segment_run_kernel(
-        mask=mask,
-        prefix_controls=prefix_controls,
-        segment_run_controls=segment_run_controls,
-        context=context,
-    ):
-        return _packed_causal_segment_run_kernel_plan(
-            segment_run_controls=segment_run_controls,
-            context=context,
-        )
-
-    if _can_use_packed_causal_segment_kernel(
-        mask=mask,
-        prefix_controls=prefix_controls,
-        segment_id_lowering=segment_id_lowering,
-        context=context,
-    ):
-        return _packed_causal_segment_kernel_plan(
-            segment_id_lowering=segment_id_lowering,
-            context=context,
-        )
-
-    return _static_splash_kernel_plan(
-        mask=mask,
-        segment_id_lowering=segment_id_lowering,
-        context=context,
-        mesh=mesh,
-    )
+    return SplashMaskPlanSpec(mask=mask_spec)
 
 
 def _unflatten_bshd(attn_output, q_class, v_class):
@@ -1662,7 +1203,7 @@ def _prepare_splash_invocation_plan(
         kv_seq_shards=kv_seq_shards,
         max_block_size=block_size,
     )
-    kernel_context = _SplashKernelContext(
+    kernel_context = SplashKernelContext(
         q_seq_len=layout.q_seq_len,
         kv_seq_len=layout.kv_seq_len,
         num_heads=layout.heads,
@@ -1675,8 +1216,8 @@ def _prepare_splash_invocation_plan(
 
     return _SplashInvocationPlan(
         mesh=mesh,
-        kernel_plan=_splash_kernel_plan(
-            mask=mask,
+        kernel_plan=prepare_splash_kernel_plan(
+            mask=_splash_mask_plan_spec_from_mask(mask),
             prefix_controls=prefix_controls,
             segment_run_controls=segment_run_controls,
             segment_id_lowering=segment_id_lowering,
